Modified test coverage config Created immutable decorator Implemented e2e test for user resolver and resolved detected bugs Generate code coverage reports

+++ backend-subli-print/src/modules/users/users.resolver.ts
@@ -1,13 +1,14 @@
 import { Resolver, Query, Mutation, Args } from '@nestjs/graphql';
-import { UsersService } from './users.service';
-import { User } from './entities/user.entity';
-import { CreateUserInput } from './dto/create-user.input';
-import { UpdateUserInput } from './dto/update-user.input';
-import { UseGuards } from '@nestjs/common';
+import { BadRequestException, UseGuards } from '@nestjs/common';
+import { isValidObjectId } from 'mongoose';
 import { GetUser } from 'src/common/decorators/get-user.decorator';
 import { UserPlainObject } from 'src/common/interfaces/user.interface';
 import { GqlJwtAuthGuard } from '../auth/guards/gql-jwt-auth.guard';
+import { UsersService } from './users.service';
+import { CreateUserInput } from './dto/create-user.input';
+import { UpdateUserInput } from './dto/update-user.input';
 import { ResetPasswordInput } from './dto/reset-pass.input';
+import { User } from './entities/user.entity';
 
 @Resolver(() => User)
 export class UsersResolver {
@@ -25,11 +26,15 @@ export class UsersResolver {
 
   @Query(() => User, { nullable: true })
   user(@Args('id', { type: () => String }) id: string) {
+    if (!id || !isValidObjectId(id))
+      throw new BadRequestException('Invalid Id');
     return this.usersService.findOne({ _id: id });
   }
 
   @Mutation(() => User, { nullable: true })
   updateUser(@Args('input') input: UpdateUserInput) {
+    if (!input.id || !isValidObjectId(input.id))
+      throw new BadRequestException('Invalid Id');
     return this.usersService.update(input);
   }
 
@@ -40,6 +45,8 @@ export class UsersResolver {
 
   @Mutation(() => Boolean)
   removeUser(@Args('id', { type: () => String }) id: string) {
+    if (!id || !isValidObjectId(id))
+      throw new BadRequestException('Invalid Id');
     return this.usersService.remove(id);
   }
